+++ src/main/java/com/christian/repository/AnimeRepositoryImpl.java
@@ -43,7 +43,7 @@ public class AnimeRepositoryImpl implements AnimeRepository {
     }
 
     @Override
-    public Anime findById(int id) {
+    public Anime findById(Long id) {
         return animeDao.findById(id);
     }
